fix(regulations): refactor regulations tests
